Add files via upload
The weather data system supports storage of temperature and humidity values from various devices, in a MongoDB database. It also has a device registry, and supports categories of users. The aim is to read device and user configuration information from config files, and populate the database with access control information, along with the actual data. The learner also needs to aggregate data periodically, and display reports based on query criteria.